feat: Initial commit of the bayes.py for the project
Initial commit of the bayes.py file for the chapter
one project of the Real World Python book.

The project covers Saving Shipwrecked Sailors with
the Bayes' Rule.

This initial commit imports the required packages,
with no errors this shows the packages are available
for python (and the project) to use.